Cancel PLAY TONE command if SIM is removed or disabled
CatService is disposed and StkAppService can also be stoped if SIM is
removed or disabled. The bad thing is that no one stops the tone being
played because of PLAY TONE command and user also has no way to stop it.
The tone must be stoped right after SIM is removed or disabled.

Bug: 132849761
Test: Confirmed the sound can be stopped with this change.

Change-Id: Iad68010a9d7ad9c72749cb88a1063de95f287588